+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 **Version 1.0.0.2016** *(TBD)*
 - Fixed a minor cosmetic bug, where switching out of the app would not always restore the correct original brightness level.
 - Added testing and QC stuff.
+- Fixed a bug, where a couple of accessibility strings in the Appearance Editor were not being properly set.
 
 **Version 1.0.0.2015** *(December 2, 2018)*
 - Tweaked the branding control in the info screen.